+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New optional `uriWithoutSessionTimeoutReset` parameter in session middleware added. There will be no session timeout reset if a request contains one of the configured paths ([PORTAL-1321](https://issue.inventage.com/browse/PORTAL-1321))
 - Add stacktrace to logback configuration ([PORTAL-1243](https://issue.inventage.com/browse/PORTAL-1243)).
 - Process exits with code 0 if any middleware initialization fails ([PORTAL-1016](https://issue.inventage.com/browse/PORTAL-1016)).
+- `uriWithoutLoggingRegex` option for RequestResponseLoggerMiddleware allowing selective logging of requests based on their URI ([PORTAL-1418](https://issue.inventage.com/browse/PORTAL-1418))
+
 ### Changed
 
 - **BREAKING**: Removed entrypoint configuration `sessionDisabled`, that was deprecated in `5.1.0` ([PORTAL-1459](https://issue.inventage.com/browse/PORTAL-1459)).